split InMemory out into published testkit module
