+++ modules/core-tests/shared/src/test/scala/SpanCoalesceTest.scala
