+++ modules/core-tests/shared/src/test/scala/SpanPropagationTest.scala
