+++ modules/testkit/shared/src/main/scala/InMemory.scala
@@ -4,13 +4,11 @@
 
 package natchez
 
-import java.net.URI
-
-import cats.data.{Chain, Kleisli}
-import cats.effect.{IO, MonadCancelThrow, Ref, Resource}
-
+import cats.data.Chain
+import cats.effect.{IO, Ref, Resource}
 import natchez.Span.Options
-import munit.CatsEffectSuite
+
+import java.net.URI
 
 object InMemory {
 
@@ -126,53 +124,4 @@ object InMemory {
     case class ReleaseRootSpan(name: String) extends NatchezCommand
   }
 
-}
-
-trait InMemorySuite extends CatsEffectSuite {
-  type Lineage = InMemory.Lineage
-  val Lineage = InMemory.Lineage
-  type NatchezCommand = InMemory.NatchezCommand
-  val NatchezCommand = InMemory.NatchezCommand
-
-  trait TraceTest {
-    def program[F[_]: MonadCancelThrow: Trace]: F[Unit]
-    def expectedHistory: List[(Lineage, NatchezCommand)]
-  }
-
-  def traceTest(name: String, tt: TraceTest) = {
-    test(s"$name - Kleisli")(
-      testTraceKleisli(tt.program[Kleisli[IO, Span[IO], *]](implicitly, _), tt.expectedHistory)
-    )
-    test(s"$name - IOLocal")(testTraceIoLocal(tt.program[IO](implicitly, _), tt.expectedHistory))
-  }
-
-  def testTraceKleisli(
-      traceProgram: Trace[Kleisli[IO, Span[IO], *]] => Kleisli[IO, Span[IO], Unit],
-      expectedHistory: List[(Lineage, NatchezCommand)]
-  ) = testTrace[Kleisli[IO, Span[IO], *]](
-    traceProgram,
-    root => IO.pure(Trace[Kleisli[IO, Span[IO], *]] -> (k => k.run(root))),
-    expectedHistory
-  )
-
-  def testTraceIoLocal(
-      traceProgram: Trace[IO] => IO[Unit],
-      expectedHistory: List[(Lineage, NatchezCommand)]
-  ) = testTrace[IO](traceProgram, Trace.ioTrace(_).map(_ -> identity), expectedHistory)
-
-  def testTrace[F[_]](
-      traceProgram: Trace[F] => F[Unit],
-      makeTraceAndResolver: Span[IO] => IO[(Trace[F], F[Unit] => IO[Unit])],
-      expectedHistory: List[(Lineage, NatchezCommand)]
-  ) =
-    InMemory.EntryPoint.create.flatMap { ep =>
-      val traced = ep.root("root").use { r =>
-        makeTraceAndResolver(r).flatMap { case (traceInstance, resolve) =>
-          resolve(traceProgram(traceInstance))
-        }
-      }
-      traced *> ep.ref.get.map { history =>
-        assertEquals(history.toList, expectedHistory)
-      }
-    }
 }
